+++ tests/extension/src/parser_func.py
@@ -1,0 +1,3 @@
+def clean(s):
+    """Custom parser for tests."""
+    return s, s.replace("-*-", "")