+++ docs/src/release_notes.rst
@@ -11,6 +11,7 @@ sphinx-codeautolink adheres to
 Unreleased
 ----------
 - Correctly test for optional types in annotations (:issue:`72`)
+- Don't check for ``notranslate`` CSS class, allow additional classes (:issue:`75`)
 
 0.7.0 (2021-11-28)
 ------------------