+++ docs/src/release_notes.rst
@@ -13,6 +13,7 @@ Unreleased
 - Correctly test for optional types in annotations (:issue:`72`)
 - Don't check for ``notranslate`` CSS class, allowing for additional classes
   (:issue:`75`)
+- Allow to specify block parsers as importable references (:issue:`76`)
 
 0.7.0 (2021-11-28)
 ------------------