+++ docs/src/release_notes.rst
@@ -15,6 +15,7 @@ Unreleased
   (:issue:`75`)
 - Allow to specify block parsers as importable references (:issue:`76`)
 - Correctly produce links for ``py`` code blocks (:issue:`81`)
+- Automatic support for ``ipython3`` code blocks (:issue:`79`)
 
 0.7.0 (2021-11-28)
 ------------------