+++ docs/src/release_notes.rst
@@ -10,6 +10,7 @@ sphinx-codeautolink adheres to
 
 Unreleased
 ----------
+- Use Sphinx logging instead of raising exceptions (:issue:`86`)
 - Use Sphinx logging instead of the builtin ``warnings`` to warn (:issue:`89`)
 
 0.8.0 (2021-12-16)