+++ docs/src/release_notes.rst
@@ -11,6 +11,7 @@ sphinx-codeautolink adheres to
 Unreleased
 ----------
 - Use Sphinx logging instead of raising exceptions (:issue:`86`)
+- Link builtins if visible to intersphinx (:issue:`87`)
 - Use Sphinx logging instead of the builtin ``warnings`` to warn (:issue:`89`)
 
 0.8.0 (2021-12-16)